requirements.txt
openai
gspread
oauth2client
python-dotenv
requests
modal
Pillow 
promptlayer